CHROMIUM: thunderbolt: Enable wakes from system suspend
In order for the router and the whole domain to wake up from system
suspend states we need to enable wakes for the connected routers. For
device routers we enable wakes from PCIe and USB 3.x. This allows
devices such as keyboards connected to USB 3.x hub that is tunneled to
wake the system up as expected. For all routers we enabled wake on USB4
for each connected ports. This is used to propagate the wake from router
to another.

Do the same for legacy routers through link controller vendor specific
registers as documented in USB4 spec chapter 13.

While there correct kernel-doc of usb4_switch_set_sleep() -- it does not
enable wakes instead there is a separate function (usb4_switch_set_wake())
that does.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:160191015
TEST=Observed no regressions with S0iX without an attached device.

TOREVERT=Once the upstream version of this patch lands in maintainer
tree, this patch should be reverted.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I36d63e50b116f8a233953d0e7ac7ac4a50aa4fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275616
Tested-by: Dana Alkattan <dana.alkattan@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>