CHROMIUM: rtw88: disable TX-AMSDU on 2.4G band
Marked as CHROMIUM because we could use can_aggregate_in_amsdu()
instead, and that's currently not in R79 4.14 kernel.

Some tests shows that using AMSDU to aggregate TCP ACKs to specific
APs will degrade the throughput on 2.4G band in 20MHz bandwidth
(< 10 Mbps, should be ~100 Mbps for 2x2). Also found that there's
barely no negative impact if we disable TX AMSDU on 2.4G to connect
to other APs. So it seems like we can just tell mac80211 to not to
aggregate MSDUs by setting the max aggregation size to 1 when
transmitting on 2.4G band.

Note that we still can TX AMSDU on 5G band and benefit from it by
having 50 ~ 70 Mbps throughput improvement.

BUG=b:124400293
BUG=b:145790461
TEST=Using RTL8822CE to connect to Cisco 1252 as AP to test T-Put,
     and see if the T-Put is severely degraded to less than 10Mbps.
     Should be around 100Mbps as RTL8822CE is 2x2 11ac device.

Change-Id: If7e8cc4157058d3e79702bfb63bf6a562e1c3790
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1958318
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2166048
Commit-Queue: Brian Norris <briannorris@chromium.org>